+++ evm_rpc_types/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - v1.2.0 Added types to support `eth_call.`
 
+### Fixed
+
+- Added missing `serde::Deserialize` annotation on `Provider`, `RpcAccess` and `RpcAuth` structs.
+
 ## [1.1.0] - 2024-10-14
 
 ### Changed